+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## future
+## v2.1.2
 ### New features
 - Readers support `Ctrl+Backspace` and `Ctrl+Delete` for deleting whole words.